feat: enable changelog generation during a release
This will generate a new changelog as part of the release, but will not
update the CHANGELOG.md (as that would move the git ref).  This means
that the CHANGELOG.md will not include the released version "at" the
version tag. Subsequent PR's will result in the changelog being updated
correctly.

Fixes #46